+++ precompiles/wasmd/tx.go
@@ -1,13 +1,11 @@
 package wasmd
 
 import (
-	"encoding/binary"
 	"fmt"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/vm"
-	"github.com/ethereum/go-ethereum/crypto"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -83,11 +81,6 @@ func (p Precompile) Execute(
 		return nil, err
 	}
 
-	// Ensure the reentrancy lock
-	if err := p.ensureLock(origin, stateDB, method); err != nil {
-		return nil, err
-	}
-
 	// Log the call
 	p.Logger(ctx).Debug(
 		"tx called",
@@ -115,54 +108,4 @@ func (p Precompile) Execute(
 
 	// Return the response
 	return method.Outputs.Pack(true)
-}
-
-// ensureLock ensures that a reentrancy lock is set and not broken for the target contract
-// Reentrance lock is built using: precompile address, origin and origin nonce
-//   - Args are avoided to build the lock key, since the attacker may manipulate it
-//
-// This is done under a transient key under Cosmos SDK's stateDB
-// The lock is released at the end of the transaction by Cosmos SDK itself
-// Locks are specifically done per transaction and not per contract:
-// - Due to a limitation between EVM and WASM gas handling
-// - And to reduce the scope of the reentrances
-// - States generated as transient states are never committed and cleared at the end of the block
-func (p Precompile) ensureLock(
-	origin common.Address,
-	stateDB vm.StateDB,
-	method *abi.Method,
-) error {
-	// Build the lock key
-	lockKey := buildReentrancyLockKey(p.Address(), origin, stateDB.GetNonce(origin))
-
-	// Check if already locked
-	if stateDB.GetTransientState(p.Address(), lockKey) == common.BytesToHash([]byte{1}) {
-		return fmt.Errorf(
-			"reentrancy detected in precompile %s, method %s",
-			p.Address().Hex(), method.Name,
-		)
-	}
-
-	// Set lock
-	stateDB.SetTransientState(p.Address(), lockKey, common.BytesToHash([]byte{1}))
-	return nil
-}
-
-// buildReentrancyLockKey builds a deterministic lock key:
-//
-//	H( "wasmd.precompile.reentrancy.lock:", precompileAddr, originAddr, originNonce[8] )
-func buildReentrancyLockKey(
-	precompileAddr common.Address,
-	origin common.Address,
-	originNonce uint64,
-) common.Hash {
-	var nonceBytes [8]byte
-	binary.BigEndian.PutUint64(nonceBytes[:], originNonce)
-
-	return crypto.Keccak256Hash(
-		reentrancyPerTargetNs,
-		precompileAddr.Bytes(),
-		origin.Bytes(),
-		nonceBytes[:],
-	)
 }