+++ wasmbinding/evm/queries.go
@@ -47,7 +47,7 @@ func (qp *QueryPlugin) HandleEVMQuery(ctx sdk.Context, evmQuery evmbindingtypes.
 			return nil, ErrExecutionReverted
 		}
 		if err != nil {
-			return nil, ErrExecutingEthCall
+			return nil, ErrExecutingEthCall.Wrap(err.Error())
 		}
 
 		bz, err := json.Marshal(res)
@@ -106,7 +106,7 @@ func (qp *QueryPlugin) HandleEthCall(ctx sdk.Context, call *evmbindingtypes.EthC
 	}
 
 	// Build the eth call request
-	req, err := buildEthCallRequest(to, data, chainID, proposer)
+	req, err := buildEthCallRequest(ctx, to, data, chainID, proposer)
 	if err != nil {
 		return nil, err
 	}
@@ -300,7 +300,7 @@ func handleRevertError(vmError string, ret []byte) error {
 }
 
 // buildEthCallRequest builds the EVM query call
-func buildEthCallRequest(to common.Address, data hexutil.Bytes, chainID *big.Int, proposer []byte) (*evmtypes.EthCallRequest, error) {
+func buildEthCallRequest(ctx sdk.Context, to common.Address, data hexutil.Bytes, chainID *big.Int, proposer []byte) (*evmtypes.EthCallRequest, error) {
 	// Build the arguments
 	args := evmtypes.TransactionArgs{
 		To:   &to,
@@ -313,10 +313,13 @@ func buildEthCallRequest(to common.Address, data hexutil.Bytes, chainID *big.Int
 		return nil, err
 	}
 
+	// Determine the gas cap for the call
+	gasCap := min(ctx.GasMeter().GasRemaining(), emvconfig.DefaultGasCap)
+
 	// Return the request
 	return &evmtypes.EthCallRequest{
 		Args:            bz,
-		GasCap:          emvconfig.DefaultGasCap,
+		GasCap:          gasCap,
 		ChainId:         chainID.Int64(),
 		ProposerAddress: sdk.ConsAddress(proposer),
 	}, nil