+++ x/feeabstraction/keeper/oracle.go
@@ -38,7 +38,7 @@ func (k Keeper) CalculateFeeTokenPrices(ctx sdk.Context) error {
 	baseTokenPrice, ok := twapPriceMap[params.NativeOracleDenom]
 	if !ok {
 		// Disable fee abstraction if there is no pricing
-		k.Logger(ctx).Debug("akii has no price, feeabstraction disabled")
+		k.Logger(ctx).Debug("%s has no price, feeabstraction disabled", params.NativeOracleDenom)
 		params.Enabled = false
 		return k.Params.Set(ctx, params)
 	}